BACKLOG-21415: use default viewMode
