+++ src/javascript/JContent/ContentRoute/ContentRoute.jsx
@@ -2,20 +2,29 @@ import React from 'react';
 import ContentLayout from './ContentLayout';
 import MainLayout from '../MainLayout';
 import ContentHeader from '../ContentHeader';
-import {ErrorBoundary, Error404, LoaderSuspense} from '@jahia/jahia-ui-root';
+import {LoaderOverlay, ErrorBoundary, Error404, LoaderSuspense} from '@jahia/jahia-ui-root';
 import {useNodeInfo} from '@jahia/data-helper';
-import {useSelector} from 'react-redux';
+import {shallowEqual, useSelector} from 'react-redux';
 import {useTranslation} from 'react-i18next';
+import JContentConstants from '~/JContent/JContent.constants';
+import {EditFrame} from '../EditFrame';
 
 export const ContentRoute = () => {
     const {t} = useTranslation('jcontent');
-    const {path, mode} = useSelector(state => ({
+    const {path, mode, viewMode} = useSelector(state => ({
         path: state.jcontent.path,
-        mode: state.jcontent.mode
-    }));
-    const res = useNodeInfo({path});
+        mode: state.jcontent.mode,
+        viewMode: state.jcontent.tableView.viewMode
+    }), shallowEqual);
 
-    if (!res.loading && (res.node === undefined || res.error)) {
+    const nodeTypes = ['jnt:page', 'jmix:mainResource'];
+    const res = useNodeInfo({path}, {getIsNodeTypes: nodeTypes});
+
+    if (res.loading) {
+        return <LoaderOverlay/>;
+    }
+
+    if (res.node === undefined || res.error) {
         if (mode === 'pages') {
             return <Error404/>;
         }
@@ -23,11 +32,14 @@ export const ContentRoute = () => {
         return <Error404 label={t('jcontent:label.contentManager.error.missingFolder')}/>;
     }
 
+    const pageBuilder = (JContentConstants.tableView.viewMode.PAGE_BUILDER === viewMode || JContentConstants.tableView.viewMode.PREVIEW === viewMode);
+    const canShowEditFrame = nodeTypes.some(nt => res.node[nt]);
+
     return (
         <MainLayout header={<ContentHeader/>}>
             <LoaderSuspense>
                 <ErrorBoundary>
-                    <ContentLayout/>
+                    { (pageBuilder && canShowEditFrame ? (<EditFrame isPreview={JContentConstants.tableView.viewMode.PREVIEW === viewMode}/>) : <ContentLayout/>) }
                 </ErrorBoundary>
             </LoaderSuspense>
         </MainLayout>
