--- src/javascript/JContent/ContentRoute/ToolBar/ViewModeSelector/ViewModeSelector.gql-queries.js
@@ -1,14 +1,0 @@
-import gql from 'graphql-tag';
-import {PredefinedFragments} from '@jahia/data-helper';
-
-export const GetContentType = gql`
-    query getContentType($path:String!) {
-        jcr {
-            node: nodeByPath(path:$path) {
-                ...NodeCacheRequiredFields
-                isDisplayableNode
-            }
-        }
-    }
-    ${PredefinedFragments.nodeCacheRequiredFields.gql}
-`;
