+++ src/javascript/JContent/EditFrame/Boxes.gql-mutations.jsx
