+++ src/javascript/JContent/EditFrame/Boxes.jsx
@@ -5,13 +5,13 @@ import {Box} from './Box';
 import {Create} from './Create';
 import PropTypes from 'prop-types';
 import {useMutation} from '@apollo/client';
-import {updateProperty} from '~/JContent/PageBuilderRoute/EditFrame/Boxes.gql-mutations';
+import {updateProperty} from '~/JContent/EditFrame/Boxes.gql-mutations';
 import {useQuery} from '@apollo/client';
-import {BoxesQuery} from '~/JContent/PageBuilderRoute/EditFrame/Boxes.gql-queries';
+import {BoxesQuery} from '~/JContent/EditFrame/Boxes.gql-queries';
 import {hasMixin, isMarkedForDeletion} from '~/JContent/JContent.utils';
-import {cmAddSelection, cmClearSelection, cmRemoveSelection} from '../../redux/selection.redux';
+import {cmAddSelection, cmClearSelection, cmRemoveSelection} from '../redux/selection.redux';
 import {batchActions} from 'redux-batched-actions';
-import {pathExistsInTree} from '../../JContent.utils';
+import {pathExistsInTree} from '../JContent.utils';
 import {useTranslation} from 'react-i18next';
 import {useNotifications} from '@jahia/react-material';
 import {refetchTypes, setRefetcher, unsetRefetcher} from '~/JContent/JContent.refetches';
