+++ src/javascript/JContent/EditFrame/Deleted.jsx
@@ -4,7 +4,7 @@ import clsx from 'clsx';
 import styles from './Deleted.scss';
 import {Delete, Typography} from '@jahia/moonstone';
 import {useTranslation} from 'react-i18next';
-import {isVisible, getCoords} from '~/JContent/PageBuilderRoute/EditFrame/EditFrame.utils';
+import {isVisible, getCoords} from '~/JContent/EditFrame/EditFrame.utils';
 
 function getBoundingBox(element) {
     if (!isVisible(element)) {
