+++ src/javascript/JContent/EditFrame/Deleted.scss
