+++ src/javascript/JContent/EditFrame/DeviceContainer.jsx
