+++ src/javascript/JContent/EditFrame/DeviceContainer.scss
