+++ src/javascript/JContent/EditFrame/EditFrame.jsx
