+++ src/javascript/JContent/EditFrame/EditFrame.scss
