+++ src/javascript/JContent/EditFrame/EditFrame.utils.jsx
