+++ src/javascript/JContent/EditFrame/Infos.jsx
