+++ src/javascript/JContent/EditFrame/LinkInterceptor.jsx
