+++ src/javascript/JContent/EditFrame/LinkInterceptor.scss
