+++ src/javascript/JContent/EditFrame/index.js
