--- src/javascript/JContent/PageBuilderRoute/index.js
@@ -1,3 +1,0 @@
-import PageBuilderRoute from './PageBuilderRoute';
-
-export default PageBuilderRoute;
